Run tests on ARM64 (#9)
The M1 or M2 chips on Apple Macs are ARM64.  [GitHub Actions do not have
a Linux ARM64 runner yet][1].  So as a workaround we use [BuildJet's
GitHub Actions ARM runners][2] to test on Linux ARM64 images
(i.e. mimicking the common setup of using a Linux devcontainer on a Mac
host computer).

We only trigger the ARM64 tests manually (i.e. on workflow_dispatch),
when we have made a change that we want to test, or when we want to do
an occasional regression test, because otherwise we would quickly use up the
BuildJet free trial minutes.

[1]: https://github.com/actions/runner-images/issues/5631
[2]: https://buildjet.com/for-github-actions/docs/runners#arm