Test Prism when container already has nodejs
Add tests to ensure that the Prism install is successful on a container
which already has node and npm installed.  NB the most useful scenario
for this is when we test on an ARM64 container, as it's ARM64 that
triggers Prism to be installed via npm rather than using a binary.