Fix Prism installation on ARM64 (#10)
* Fix Prism installation on ARM64

Prior to this commit the Prism feature was failing to install Prism on
ARM64 (e.g. when the host machine was a Mac with an M1 or M2 chip). This
is because [currently the Prism maintainers do not currently publish an
ARM64 binary as part of their release process][1].  The fix in this
commit is to fallback to installing via npm if on ARM64.

To facilitate this the feature now installs node and npm if they are not
already installed on the container.  In theory we could ensure the
feature is completely idempotent by packaging a binary and then
uninstalling node and npm if they were installed by the feature, but
that would be overkill for now at least.  We can always do this in a
future PR if there is a need for it.

[1]: https://github.com/stoplightio/prism/issues/2055

* Test Prism when container already has nodejs

Add tests to ensure that the Prism install is successful on a container
which already has node and npm installed.  NB the most useful scenario
for this is when we test on an ARM64 container, as it's ARM64 that
triggers Prism to be installed via npm rather than using a binary.